Initial commit after fork from origin repo

+++ API/DTOs/StanjePacijentaReadDTO.cs
@@ -13,6 +13,6 @@ namespace API.DTOs
         public int Kasalj { get; set; }
         public int Umor { get; set; }
         public int BolUMisicima { get; set; }
-        public int Vrijeme { get; set; }
+        public long Vrijeme { get; set; }
     }
 }
