+++ API/EFModels/_02StanjePacijenta.cs
@@ -11,7 +11,7 @@ namespace API.EFModels
         public int Kasalj { get; set; }
         public int Umor { get; set; }
         public int BolUMisicima { get; set; }
-        public int Vrijeme { get; set; }
+        public long Vrijeme { get; set; }
 
         public virtual _02Pacijent Korisnik { get; set; }
     }
