+++ API/Models/StanjePacijenta.cs
@@ -8,6 +8,6 @@
         public int Kasalj { get; set; }
         public int Umor { get; set; }
         public int BolUMisicima { get; set; }
-        public int Vrijeme { get; set; }
+        public long Vrijeme { get; set; }
     }
 }
